+++ src/main/java/yoda/task/DeadlineTask.java
@@ -1,29 +1,29 @@
+package yoda.task;
+
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 import java.util.Locale;
 
-public class EventTask extends Task {
-    private final String startPretty;
-    private final String endPretty;
-    private final String startIso;
-    private final String endIso;
+public class DeadlineTask extends Task {
+    private final String deadlinePretty;
+    private final String deadlineIso;
 
     private static final DateTimeFormatter OUT_DATE =
             DateTimeFormatter.ofPattern("MMM d yyyy", Locale.ENGLISH);
     private static final DateTimeFormatter OUT_DATETIME =
             DateTimeFormatter.ofPattern("MMM d yyyy, HH:mm", Locale.ENGLISH);
 
-    //for date-only inputs
+    // accepted date-only inputs
     private static final DateTimeFormatter[] DATE_PATTERNS = new DateTimeFormatter[] {
             DateTimeFormatter.ISO_LOCAL_DATE,               // 2019-12-02
             DateTimeFormatter.ofPattern("d/M/uuuu"),        // 2/12/2019
             DateTimeFormatter.ofPattern("d-M-uuuu"),        // 2-12-2019
-            DateTimeFormatter.ofPattern("MMM d uuuu", Locale.ENGLISH)
+            DateTimeFormatter.ofPattern("MMM d uuuu", Locale.ENGLISH) // Dec 2 2019 (for old saved data, optional)
     };
 
-    //for date+time inputs
+    // accepted date+time inputs
     private static final DateTimeFormatter[] DATETIME_PATTERNS = new DateTimeFormatter[] {
             DateTimeFormatter.ISO_LOCAL_DATE_TIME,          // 2019-12-02T18:00
             DateTimeFormatter.ofPattern("d/M/uuuu HHmm"),   // 2/12/2019 1800
@@ -33,55 +33,33 @@ public class EventTask extends Task {
             DateTimeFormatter.ofPattern("MMM d uuuu, HH:mm", Locale.ENGLISH) // Dec 2 2019, 18:00
     };
 
-    public EventTask(String desc, String startInput, String endInput) {
+    public DeadlineTask(String desc, String deadlineInput) {
         super(desc);
 
-        String[] s = parseAny(startInput);
-        this.startPretty = s[0];
-        this.startIso = s[1];
-
-        String[] e = parseAny(endInput);
-        this.endPretty = e[0];
-        this.endIso = e[1];
-    }
-
-    @Override
-    public String toString(){
-        return "[E]" + "[" + this.getStatusIcon() + "] " + this.description +
-                " (from: " + this.startPretty + " to: " + this.endPretty + ")";
-    }
-
-    @Override
-    public String toSaveLine() {
-        return "E | " + (isDone ? 1 : 0) + " | " + description + " | " + startIso + " | " + endIso;
-    }
-
-    //helper functions
-
-    // returns {pretty, iso}; throws if cannot parse
-    private static String[] parseAny(String input) {
-        if (input == null || input.isBlank()) {
-            throw new IllegalArgumentException("Description, start and end, please tell");
+        if (deadlineInput == null || deadlineInput.isBlank()) {
+            this.deadlinePretty = "—";
+            this.deadlineIso = "";
+            return;
         }
 
-        LocalDateTime ldt = tryParseDateTime(input);
+        LocalDateTime ldt = tryParseDateTime(deadlineInput);
         if (ldt != null) {
-            return new String[] {
-                    ldt.format(OUT_DATETIME),
-                    ldt.format(DateTimeFormatter.ISO_LOCAL_DATE_TIME)   // 2019-12-02T18:00
-            };
+            this.deadlinePretty = ldt.format(OUT_DATETIME);
+            this.deadlineIso = ldt.format(DateTimeFormatter.ISO_LOCAL_DATE_TIME); // 2019-12-02T18:00
+            return;
         }
 
-        LocalDate ld = tryParseDate(input);
+        LocalDate ld = tryParseDate(deadlineInput);
         if (ld != null) {
-            return new String[] {
-                    ld.format(OUT_DATE),
-                    ld.format(DateTimeFormatter.ISO_LOCAL_DATE)         // 2019-12-02
-            };
+            this.deadlinePretty = ld.format(OUT_DATE);
+            this.deadlineIso = ld.format(DateTimeFormatter.ISO_LOCAL_DATE); // 2019-12-02
+            return;
         }
 
-        throw new IllegalArgumentException("Description, start and end, please tell");
+        throw new IllegalArgumentException("Description and deadline, please tell");
     }
+
+    public DeadlineTask(String desc) { this(desc, ""); }
 
     private static LocalDateTime tryParseDateTime(String s) {
         for (DateTimeFormatter f : DATETIME_PATTERNS) {
@@ -95,5 +73,15 @@ public class EventTask extends Task {
             try { return LocalDate.parse(s, f); } catch (DateTimeParseException ignored) {}
         }
         return null;
+    }
+
+    @Override
+    public String toString(){
+        return "[D][" + this.getStatusIcon() + "] " + this.description + " (by: " + this.deadlinePretty + ")";
+    }
+
+    @Override
+    public String toSaveLine() {
+        return "D | " + (isDone ? 1 : 0) + " | " + description + " | " + deadlineIso;
     }
 }
