+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 _______________________
-Yoda, I am.
+yoda.Yoda, I am.
 _______________________
 Do for you, what can I?
